Fixed the ir problem.
It turns out to be a firmware problem. The arduino communication code
added 1 to bytes to avoid sending null bytes, but the python code on the
other side failed to subtract 1 from the bytes. This makes a big
difference for the analog inputs because each input has two bytes (and a
difference of 1 in the large byte is a significant change).

Separately, we discovered that I had set the IR distance cutoff at 40cm,
where it starts to deviate from the model. We changed it so it now reads
distances up to 80 cm (it is only accurate up to 40cm, but it will
continue to give increasing values for larger distances up to 80cm
(which it displays as 155cm)).

avoid sending null bytes but failing to subtract