Simplified ball management algorithm.
Instead of remembering balls and matching newly found balls with old
balls, it is now more of a storage location for vision results (it
always targets the closest ball; it only knows about balls that were
seen in the most recent frame).

This version of the algorithm is probably sufficient (although not as
good), but will be much more reliable and easier to get working by
competition time.