Wrote code for wall detection and wall scoring.
Obviously none of it has been tested on the robot. But wall detection
has been tested in cvTest on reference images and seems to work. It runs
quickly as compared to ball detection, but hopefully both are fast
enough. It would really suck to have to optimize them now.

The procedure for scoring balls is
1) approach yellow wall, keeping it centered in view
2) when you get close-ish to the wall, calculate the angle between the
robot and the wall using the IR sensors and use PID to turn
perpendicular (ish) to the wall (still advancing, but more slowly)
3) when a bump sensor triggers, stop, lift the balls, drop the ramp, and
celebrate. Perhaps it needs to play a sound file here?