Fixed more bugs; preliminary code runs without errors.
Still hasn't been run with a robot, but it has been run with an arudino
(no sensors, just a start button) and a camera. Without errors.

The image processing was really slow but I fixed it by shrinking the
image in ImageAcquisition (rather than ImageProcessing). It's still
slower than it should be, but it is fast enough to work. Later I'll
have to figure out why it is slow.